+++ internal/githubclient/githubclient.go
@@ -10,7 +10,7 @@ import (
 	"net/url"
 	"os"
 
-	"github.com/google/go-github/github"
+	"github.com/google/go-github/v28/github"
 	"golang.org/x/oauth2"
 )