+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 Fixed:
 
+- default mods overrode loaded ones
 - archive:changelog to cut only the last release announcement
 
 ## [v0.1.0] - 2019-11-16