+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+Changed:
+
+- upgrade to go-github/v28
+
 Fixed:
 
 - default mods overrode loaded ones