+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+No changes so far.
+
+## [v0.2.0] - 2019-11-17
+
 Added:
 
 - variable expansion for modules.TemplateData.Parse
@@ -43,5 +47,6 @@ Changed:
 - rename project to goshipdone
 - fold release_note into archive
 
-[Unreleased]: https://github.com/julian7/goshipdone/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/julian7/goshipdone/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/julian7/goshipdone/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/julian7/goshipdone/releases/tag/v0.1.0